Import from index. add getFlowBalance helper
