+++ lib/javascript/templates/scripts/get_balance.cdc
@@ -1,0 +1,12 @@
+// This script reads the balance field of an account's FlowToken Balance
+
+import FungibleToken from 0xFUNGIBLETOKENADDRESS
+import ExampleToken from 0xTOKENADDRESS
+
+pub fun main(account: Address): UFix64 {
+    let acct = getAccount(account)
+    let vaultRef = acct.getCapability(/public/exampleTokenBalance)!.borrow<&ExampleToken.Vault{FungibleToken.Balance}>()
+        ?? panic("Could not borrow Balance reference to the Vault")
+
+    return vaultRef.balance
+}
